Narrow CharsXY to u16 (from usize)
Using usize to represent character screen coordinates was overkill, and
especially now that the graphical coordinates are capped at i16.  To
make things more homogeneous, cap CharsXY to u16.

Note that PixelsXY allows negative numbers to support partial off-screen
rendering, which is very useful, but CharsXY does not.  The reason is
that the console primitives do not allow off-screen printing and thus
we shouldn't be worrying about negative numbers everywhere.

+++ std/src/console/mod.rs
@@ -99,15 +99,15 @@ pub enum ClearType {
 #[derive(Clone, Copy, Debug, Default, Eq, PartialEq)]
 pub struct CharsXY {
     /// The column number, starting from zero.
-    pub x: usize,
+    pub x: u16,
 
     /// The row number, starting from zero.
-    pub y: usize,
+    pub y: u16,
 }
 
 impl CharsXY {
     /// Constructs a new coordinate at the given `(x, y)` position.
-    pub fn new(x: usize, y: usize) -> Self {
+    pub fn new(x: u16, y: u16) -> Self {
         Self { x, y }
     }
 }
